Bugfix for custom color array size
Probably a bug occurred, when replacing hard wired 64 by CHERRY_KB_MAX_PACKET_SIZE. In "DeviceUpdateLEDs()", the former value was 54 * 7 (not 64 * 7). Note that 54*7 = 378, but it would seem more logical to use 63*6 which is also 378 and refers to the six rows of the array. That would correspond to 21*6*3, so 21 columns times 6 rows times 3 color components. However, I found that it's best to let the array dimensions correspond to the matrix_map[6][23] which is defined on top of the cpp file, because storing the custom colors to a profile and then reloading the profile yields the correct result. It is possible that the dimensions [6][21] should be used for both arrays, but currently there is no evident reason to do so.