Mode switch repair for FL_10.0N by adding usb_buf[0x07] = 0x55;
I observed in USB captured packages, that the supposed "0x06 packages" seems to be actually a "0x08 package" and needs the appendix of one zero byte and 0x55 as the last byte. That seems to be irrelevant for the TKL RGB, but on the "MX Board 10.0N FL RGB", mode switching will work only with 0x55 as last byte (on index 0x07).